Switch `u64` to `usize`

+++ compiler+runtime/src/cpp/jank/runtime/core/array.cpp
@@ -12,12 +12,12 @@ namespace jank::runtime
     return try_object<obj::array<int>>(o)->canonical_type;
   }
 
-  obj::bool_array_ref boolean_array(u64 const size)
+  obj::bool_array_ref boolean_array(usize const size)
   {
     return make_box<obj::array<bool>>("bool", size);
   }
 
-  obj::bool_array_ref boolean_array(u64 const size, bool const init_value)
+  obj::bool_array_ref boolean_array(usize const size, bool const init_value)
   {
     return make_box<obj::array<bool>>("bool", size, init_value);
   }
@@ -41,12 +41,12 @@ namespace jank::runtime
     return a;
   }
 
-  obj::char_array_ref char_array(u64 const size)
+  obj::char_array_ref char_array(usize const size)
   {
     return make_box<obj::array<char>>("char", size);
   }
 
-  obj::char_array_ref char_array(u64 const size, char const init)
+  obj::char_array_ref char_array(usize const size, char const init)
   {
     return make_box<obj::array<char>>("char", size, init);
   }
@@ -88,12 +88,12 @@ namespace jank::runtime
     return a;
   }
 
-  obj::f32_array_ref float_array(u64 const size)
+  obj::f32_array_ref float_array(usize const size)
   {
     return make_box<obj::array<f32>>("f32", size);
   }
 
-  obj::f32_array_ref float_array(u64 const size, f32 const init)
+  obj::f32_array_ref float_array(usize const size, f32 const init)
   {
     return make_box<obj::array<f32>>("f32", size, init);
   }
@@ -117,12 +117,12 @@ namespace jank::runtime
     return a;
   }
 
-  obj::f64_array_ref double_array(u64 const size)
+  obj::f64_array_ref double_array(usize const size)
   {
     return make_box<obj::array<f64>>("f64", size);
   }
 
-  obj::f64_array_ref double_array(u64 const size, f64 const init)
+  obj::f64_array_ref double_array(usize const size, f64 const init)
   {
     return make_box<obj::array<f64>>("f64", size, init);
   }
@@ -146,12 +146,12 @@ namespace jank::runtime
     return a;
   }
 
-  obj::i16_array_ref short_array(u64 const size)
+  obj::i16_array_ref short_array(usize const size)
   {
     return make_box<obj::array<i16>>("i16", size);
   }
 
-  obj::i16_array_ref short_array(u64 const size, i16 const init)
+  obj::i16_array_ref short_array(usize const size, i16 const init)
   {
     return make_box<obj::array<i16>>("i16", size, init);
   }
@@ -175,12 +175,12 @@ namespace jank::runtime
     return a;
   }
 
-  obj::i32_array_ref int_array(u64 const size)
+  obj::i32_array_ref int_array(usize const size)
   {
     return make_box<obj::array<i32>>("i32", size);
   }
 
-  obj::i32_array_ref int_array(u64 const size, i32 const init)
+  obj::i32_array_ref int_array(usize const size, i32 const init)
   {
     return make_box<obj::array<i32>>("i32", size, init);
   }
@@ -204,12 +204,12 @@ namespace jank::runtime
     return a;
   }
 
-  obj::i64_array_ref long_array(u64 const size)
+  obj::i64_array_ref long_array(usize const size)
   {
     return make_box<obj::array<i64>>("i64", size);
   }
 
-  obj::i64_array_ref long_array(u64 const size, i64 const init)
+  obj::i64_array_ref long_array(usize const size, i64 const init)
   {
     return make_box<obj::array<i64>>("i64", size, init);
   }
@@ -233,7 +233,7 @@ namespace jank::runtime
     return a;
   }
 
-  obj::object_array_ref object_array(u64 const size)
+  obj::object_array_ref object_array(usize const size)
   {
     return make_box<obj::array<object_ref>>("object", size);
   }
@@ -257,12 +257,12 @@ namespace jank::runtime
     return a;
   }
 
-  obj::u8_array_ref byte_array(u64 const size)
+  obj::u8_array_ref byte_array(usize const size)
   {
     return make_box<obj::array<u8>>("u8", size);
   }
 
-  obj::u8_array_ref byte_array(u64 const size, u8 const init)
+  obj::u8_array_ref byte_array(usize const size, u8 const init)
   {
     return make_box<obj::array<u8>>("u8", size, init);
   }
